Breaking changes & retaining some junk that would be cleaned

+++ stop-embedded-etcd.sh
@@ -1,7 +1,17 @@
 #!/usr/bin/env bash
 . utils.sh
+
 if [ "$this_host_ip" = $master_ip ]; then
-  sudo rm /etc/kubernetes/manifests/etcd.yaml
+  if [ -f /etc/kubernetes/manifests/etcd.yaml ]; then
+    sudo rm /etc/kubernetes/manifests/etcd.yaml
+  else
+    err "Stopping etcd - no etcd.yaml @$this_host_ip"
+  fi
 else
-  sudo -u $usr ssh -o "StrictHostKeyChecking no" -o "ConnectTimeout=5" $master_ip rm -f /etc/kubernetes/manifests/etcd.yaml
+  . execute-command-remote.sh $master_ip ls /etc/kubernetes/manifests/etcd.yaml
+  if [ "$?" -eq 0 ]; then
+    . excute-command-remote.sh $master_ip rm /etc/kubernetes/manifests/etcd.yaml
+  else
+    err "Stopping etcd - no etcd.yaml @$master_ip"
+  fi
 fi
